Even more aggressive backend refactoring and apply known fixes to dawn backends
