+++ src/gt4py/backend/__init__.py
@@ -18,9 +18,7 @@ from .base import *
 
 from .debug_backend import *
 from .numpy_backend import *
-from .base_gt_backend import *
-from .gt_cpu_backend import *
-from .gt_cuda_backend import *
-from .dawn_backend import *
+from .gt_backends import *
+from .dawn_backends import *
 
 from . import python_generator
